requirements.txt
# Core Jupyter dependencies
jupyter
jupyter-cache
nbclient
nbformat
python-dotenv

# Plotting and data
matplotlib
numpy
pandas

# DSPy tutorial dependencies
dspy-ai
attachments
datar
fastcore
claudette
